refactor(dpos)!: take the staking ABI and its limits from the shared crates
The node's four independent sol! blocks are gone — node/src/evm.rs,
staking-reader/reader.rs, node/src/slasher_sink.rs and
consensus/slasher/actor.rs all import fluentbase-staking-abi, the same
declaration the contract derives its dispatch selectors from. Every
"MUST mirror" constant under crates/dpos is now a pub use of
fluentbase_types::staking_protocol, and the ahead-commit horizon reads
MAX_COMMITTEE_LOOKAHEAD_EPOCHS instead of a literal 2.

Two real disagreements surfaced before any of it was moved.

getEpochBlockInterval, getActiveValidatorsLength and getUndelegatePeriod
return uint64 on chain (storage.rs holds all three as StorageU64) and were
declared uint32 here. Resolved toward the contract, which widens
epoch_block_interval to u64 through epoch_transition, dpos.rs, cert_inlet
and outer.rs; outer.rs loses the debug_assert that guarded a NonZeroU64→u32
narrowing that no longer exists. No selector moves — return types do not
enter one.

epoch_of_block and the contract's epoch_at_block differed in ARGUMENT ORDER
and at activation == 0. The formula and the pre-activation clamp are now
staking_protocol::epoch_at_block, shared; the "activation 0 is the unarmed
sentinel" arm stays contract-side, because it is a rule about a state only
the contract is in and the node's scheduled_dpos_activation folds a zero
before any epoch is computed. epoch_of_block is deleted.

Two node comments were false and are deleted with the code that carried
them: "the contract has no slashEquivocation counterpart at all" (it does)
and "KNOWN CONTRACT DRIFT — the fourth array has no contract-side
counterpart" (it returns four). So is the 110-line CONTRACT ABI DELTA —
MERGE CHECKLIST in slasher/actor.rs: the branch it described merged as
f16fdd90 and its premise is not in this tree.

Deleted tests, each of which compared one declaration with itself:
view_selectors_are_pinned, epoch_committee_return_arity_is_pinned,
commit_epoch_committee_selector_is_pinned,
next_epoch_to_commit_selector_is_pinned, and the signature-literal halves
of the calldata and close-event pins. What replaces them is one pin per
side against an external witness — the ABI crate's
selectors_match_the_deployed_artefact_scan against the blob scan, the
contract's derived_selectors_match_independent_hex_pins — plus
the_shared_fault_budget_is_commonwares_budget, which checks the shared
fault_tolerance against commonware's N3f1::max_faults over the whole legal
committee range instead of against a second copy of (n-1)/3. The cast
calldata and cast abi-encode vectors stay: those witnesses are real.

Blob rebuilt and re-vendored; digests and the widened selector scan
recorded in STAKING_ARTEFACT.md, whose provenance header now says the
sources live in this tree. Live: make case-growth PASS on the rebuilt
image, committee 4→5→6 across two boundaries, with
epoch_committee_committed decoding on every node — which is the shared
selector and the shared event topic0 agreeing on a real chain.

BREAKING CHANGE: RethStakingStateReader::epoch_block_interval and
active_validators_length return u64; reader::epoch_of_block is replaced by
reader::epoch_at_block, which takes (block, activation, interval) and
returns Option<u64>.

+++ crates/dpos/consensus/tests/slasher_integration.rs
@@ -216,7 +216,7 @@ impl StakingStateRead for StubReader {
             Ok(self.snapshot.clone())
         }
     }
-    fn epoch_block_interval(&self, _at: B256) -> Result<u32, ReadError> {
+    fn epoch_block_interval(&self, _at: B256) -> Result<u64, ReadError> {
         Ok(100)
     }
     fn dpos_activation_block(&self, _at: B256) -> Result<u64, ReadError> {
